Extracts fake data helper functions into separate file. Index.js now just contains the functions for seeding the database, as well as a select query for pulling all relevant data for a given recipe id
